Implement VFS, COMMAND.COM terminal, and File Viewer
Add a virtual file system with drive-letter addressing (A:/B: scratch in
IndexedDB, C: per-session shared, U: per-user private), a COMMAND.COM
terminal emulator with DOS-style commands (DIR, COPY, DEL, CD, TYPE,
ATTRIB, GET, PUT, etc.), and a dual-pane File Viewer with multi-select,
Ctrl/Shift-click, location bar navigation, and upload/download support.

Core VFS features:
- Diesel migration for vfs_files table with drive-scoped unique indexes
- Path parser with drive letter + Unix-style path normalization
- CRUD operations: read, write, list, stat, copy, rename, delete, mkdir
- Inline storage (<=8KB) vs CAS reference for large files
- Unix-style permissions (rwx, umask, chmod) with GM override
- Fnmatch glob pattern matching (*, ?) for wildcard commands
- Quota enforcement per drive (10MB scratch, 100MB session, 10/20MB user)
- VfsChanged WebSocket broadcast for C: drive modifications
- 79 tests covering paths, DB ops, permissions, fnmatch, and copy

Terminal (COMMAND.COM):
- Client-side WASM shell with command parsing and history
- Drive letter switching, EXIT command, working directory tracking
- FTP-style GET/PUT for browser file transfer
- Shared browser_helpers module for file picker and upload

File Viewer:
- Dual-pane browser with independent navigation per pane
- Multi-select with Ctrl+click and Shift+click
- Toolbar: copy, delete, mkdir, upload, download
- Drive list with usage/quota display
- Help page with keyboard shortcuts

# Feature 34: Virtual File System — CORE DONE (scratch drives, archive pending)
# Feature 36: VFS Terminal Shell — DONE
# Feature 37: VFS File Browser — CORE DONE (drag-drop, zip pending)

+++ src/server/media_handler.rs
@@ -123,7 +123,7 @@ pub async fn upload_media(
             hash: &hash,
             content_type: &content_type,
             media_type,
-            size_bytes: bytes.len() as i32,
+            size_bytes: bytes.len() as i64,
             uploaded_by: user_id,
         };
